Make tos_warning() an (automatically run) method of ISI
Since ISI() is only called if there's actually connections to be made,
this actually makes the scraper quieter, which is nice. You might balk at
putting prints inside an __init__, but I think it is fair:
it only warns you if it's actually doing stuff to warn you about,
and whenever you are using ISI() you *should* be warned.